api/requirements.txt
fastapi==0.95.2
pydantic==1.10.7
uvicorn==0.22.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-multipart==0.0.6
pydub==0.25.1
